convert img URLs to be protocol-relative
Added a new filter to the markdown processing pipeline to convert image
URLs to use a protocol that matches the protocol being used to view the
current page. This will "convert" images linked to by, say, HTTP to be
viewed over HTTPS and not trigger a security alert on the page.

Signed-off-by: Robb Kidd <rkidd@chef.io>