The FIPSification of Supermarket
Three of the five services that comprise Supermarket are identified as
appropriate for enabling FIPS mode on:

* front-facing reverse web proxy (nginx): in single-server
  standalone environments nginx is responsible for SSL termination
* web application service (rails): some crypto is performed by the web
  service for user authentication via hashing secrets
* background jobs (sidekiq): no crypto for secrets is performed in
  background jobs as of this writing, but the jobs share a code base
  and runtime configuration with the web service, so enabling it is no
  more difficult

+ new `['supermarket']['fips_enabled']` item added to the omnibus install
  to explicitly enable or disable FIPS mode; the default is set to nil
  which the omnibus install will treat as "determine FIPS mode based on
  whether FIPS is enabled in the kernel"
+ new `['supermarket']['ssl']['fips_ciphers']` added to provide
  overridable default ciphers when operating in FIPS mode

The service startup scripts for nginx, rails, and sidekiq are updated to
`export OPENSSL_FIPS=1` per the OpenSSL documentation for enabling FIPS
mode when `['supermarket']['fips_enabled']` is true--whether explicitly
or when detected as enabled in the kernel.

An initializer and a library have been added to the web application so
that the rails and sidekiq services react to this environment variable
by enabling `fips_mode` in the loaded Ruby OpenSSL library and by using
the hash algorithms defined in the OpenSSL library's FIPS object module.

Use the latest version of the OpenSSL FIPS object module, as of this
writing version 2.0.16.

The `install-check` Inspec tests are updated with a FIPS section to test
that an environment variable `OPENSSL_FIPS` has been set to `1` for the
three services and that FIPS ciphers are set in the nginx SSL
configuration. The supermarket-ctl test command is updated to allow
skips so that the FIPS check can be skipped on non-FIPS installs.

Signed-off-by: Robb Kidd <rkidd@chef.io>